# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: StreamGate API
  description: |
    NFT-gated streaming platform API. Core flow: wallet sign-in → NFT verification → protected HLS streaming.

    ## Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```

    ## NFT Permission Control
    Content is protected by NFT ownership. Users must hold the specified NFT to access protected streaming content.
  version: 1.0.0
  contact:
    name: StreamGate Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Relative path (behind proxy)
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /upload:
    post:
      tags: [Upload]
      summary: Upload a whole file
      description: Uploads a complete file in a single request
      operationId: uploadFile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: File uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadResponse"
  /upload/list:
    get:
      tags: [Upload]
      summary: List uploads
      description: Lists upload records for the authenticated user
      operationId: listUploads
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Upload list
  /upload/init:
    post:
      tags: [Upload]
      summary: Initialize chunked upload
      description: Starts a chunked upload session
      operationId: initChunkedUpload
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateUploadRequest"
      responses:
        "200":
          description: Upload session initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitiateUploadResponse"
  /upload/chunk:
    post:
      tags: [Upload]
      summary: Upload a file chunk
      description: Uploads a single chunk in a chunked upload session
      operationId: uploadChunk
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, upload_id, chunk_index]
              properties:
                file:
                  type: string
                  format: binary
                upload_id:
                  type: string
                chunk_index:
                  type: integer
      responses:
        "200":
          description: Chunk received
  /upload/{id}/complete:
    post:
      tags: [Upload]
      summary: Complete chunked upload
      description: Finalizes a chunked upload session and triggers transcoding
      operationId: completeChunkedUpload
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Upload session ID
      responses:
        "200":
          description: Upload completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteUploadResponse"
  /upload/{id}/complete-upload:
    post:
      tags: [Upload]
      summary: Complete upload (alternative)
      description: Alternative endpoint to finalize an upload
      operationId: completeUpload
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upload completed
  /upload/{id}/status:
    get:
      tags: [Upload]
      summary: Get upload status
      operationId: getUploadStatus
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Upload status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadStatus"
        "404":
          description: Upload session not found
  /upload/{id}/download-url:
    get:
      tags: [Upload]
      summary: Get download URL
      description: Returns a presigned download URL for the uploaded file
      operationId: getDownloadURL
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Download URL
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    InitiateUploadRequest:
      type: object
      required: [filename, file_size, chunk_size]
      properties:
        filename:
          type: string
          description: Name of the file
        file_size:
          type: integer
          format: int64
          description: Size of the file in bytes
        chunk_size:
          type: integer
          description: Size of each chunk in bytes
          default: 5242880
    InitiateUploadResponse:
      type: object
      properties:
        upload_id:
          type: string
          description: Unique upload session ID
        chunk_size:
          type: integer
        total_chunks:
          type: integer
        expires_at:
          type: string
          format: date-time
    UploadResponse:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
    CompleteUploadResponse:
      type: object
      properties:
        file_id:
          type: string
        upload_id:
          type: string
        file_size:
          type: integer
          format: int64
        transcoding_job_id:
          type: string
    UploadStatus:
      type: object
      properties:
        upload_id:
          type: string
        status:
          type: string
          enum: [pending, uploading, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        uploaded_bytes:
          type: integer
          format: int64
        total_bytes:
          type: integer
          format: int64
        uploaded_chunks:
          type: integer
        total_chunks:
          type: integer
...
security:
  - bearerAuth: []